+++ src/signposting/htmllinks.py
@@ -25,7 +25,7 @@ from bs4 import BeautifulSoup,SoupStrainer
 from .signpost import SIGNPOSTING,AbsoluteURI,Signpost,Signposting
 
 def find_signposting_html(uri:Union[AbsoluteURI, str]) -> Signposting:
-    """Parse HTML to find <link> elements for signposting.
+    """Parse HTML to find ``<link>`` elements for signposting.
     
     HTTP redirects will be followed and any relative paths in links
     made absolute correspondingly.
@@ -36,7 +36,7 @@ def find_signposting_html(uri:Union[AbsoluteURI, str]) -> Signposting:
     :throws requests.HTTPError: If the HTTP request failed, e.g. 404 Not Found
     :throws UnrecognizedContentType: If the HTTP resource was not a recognized HTML/XHTML content type
     :throws HTMLParser.HTMLParseError: If the HTML could not be parsed.
-    :returns: A parsed `Signposting` object (which may be empty)
+    :returns: A parsed :class:`Signposting` object (which may be empty)
     """
     html = _get_html(AbsoluteURI(uri))
     return _parse_html(html)