+++ src/signposting/signpost.py
@@ -31,34 +31,30 @@ validity at construction time, so that consumers of
 about type safety.
 """
 
+# PEP-563 support self-references of types in class definitions
+from __future__ import annotations
+
 import itertools
-from multiprocessing import AuthenticationError
 import re
-from types import NoneType
 from typing import Collection, Iterable, Iterator, List, Optional, Set, Sized, Tuple, Union, AbstractSet, FrozenSet
 from enum import Enum, auto, unique
-import warnings
+from warnings import warn
 
 import rfc3987
 import urllib.parse
 from urllib.parse import urljoin
 from httplink import Link
-from warnings import warn
-
 
 class AbsoluteURI(str):
     """An absolute URI, e.g. "http://example.com/" """
     def __new__(cls, value: str, base: Optional[str] = None):
-        """Create an absolute URI reference.
+        """Create and validate an absolute URI reference.
 
-        If the base parameter is given, it is used to resolve the
-        potentially relative URI reference, otherwise the first argument
-        must be an absolute URI.
+        Note that IRIs are not supported unless ``%``-encoded.
 
-        This constructor will throw `ValueError` if the
-        final URI reference is invalid or not absolute.
-
-        Note that IRIs are not supported.
+        :param value: URI string to validate as Absolute URI. May be a relative URI reference if `base` is provided.
+        :param base: (Optional) URI used to resolve the potentially relative URI reference, otherwise `value` must be an absolute URI.
+        :raise ValueError: if the final URI reference is invalid or not absolute.
         """
         if isinstance(value, cls):
             return value # Already AbsoluteURI, no need to check again
@@ -81,15 +77,15 @@ class AbsoluteURI(str):
 
 
 class MediaType(str):
-    """An IANA media type, e.g. text/plain.
+    """An IANA media type, e.g. ``text/plain``.
 
     This class ensures the type string is valid according to `RFC6838`_
     and for convenience converts it to lowercase.
 
     While the constructor do check that the main type is an offical IANA subtree
-    (see `MediaType.MAIN`), it does not enforce the individual subtype to be registered.
+    (see :attr:`MAIN`), it does not enforce the individual subtype to be registered.
     In particular RFC6838 permits unregistered subtypes
-    starting with `vnd.`, `prs.` and `x.`
+    starting with ``vnd.``, ``prs.`` and ``x.``
 
     Extra content type parameters such as ``;profile=http://example.com/`` are
     **not** supported by this class, as they do not form part of the
@@ -115,10 +111,10 @@ class MediaType(str):
     """
 
     main: str
-    """The main type, e.g. image"""
+    """The main type, e.g. ``image``"""
 
     sub: str
-    """The sub-type, e.g. jpeg"""
+    """The sub-type, e.g. ``jpeg``"""
 
     def __new__(cls, value=str):
         """Construct a MediaType.
@@ -157,8 +153,8 @@ class LinkRel(str, Enum):
     only link relations listed in `FAIR`_ and `signposting`_
     conventions are included in this enumerator.
 
-    A link relation enum can be looked up from its RFC8288 _value_
-    by calling ``LinkRel("cite-as")`` - note that this particular
+    A link relation enum can be looked up from its RFC8288 *value*
+    by calling ``LinkRel("cite-as")`` -- note that this particular
     example has a different Python-compatible spelling in it's
     enum *name* (``LinkRel.cite_as``).
 
@@ -192,7 +188,7 @@ class Signpost:
     or otherwise constructed.
 
     In some case the link relation may have additional attributes,
-    e.g. ``signpost.link["title"]`` - the purpose of this class is however to
+    e.g. ``signpost.link["title"]`` -- the purpose of this class is however to
     lift only the navigational attributes for FAIR Signposting.
     """
 
@@ -212,7 +208,7 @@ class Signpost:
     retrieving the target URI.
 
     This property is optional, and should only be expected
-    if `rel` is :const:`LinkRel.describedby` or :const:`LinkRel.item`
+    if ``rel`` is :const:`LinkRel.describedby` or :const:`LinkRel.item`
     """
 
     # FIXME: Correct JSON-LD profile
@@ -235,16 +231,16 @@ class Signpost:
     from the one originally requested.
 
     This attribute is optional (with ``None`` indicating unknown context),
-    however producers of ``Signpost`` instances from are encouraged to 
-
+    context may be implied from the resource, e.g. as indicated by 
+    :attr:`Signposting.context_url`
     """
 
     link: Optional[Link]
-    """The Link object this signpost was created from.
+    """The :class:`Link` object this signpost was created from.
 
     May contain additional attributes such as ``link["title"]``.
-    Note that a single Link may have multiple ``rel``s, therefore it is
-    possible that multiple :class:`Signpost`s refer to the same link.
+    Note that a single Link may have multiple ``rel`` relations, therefore it is
+    possible that multiple :class:`Signpost` instances refer to the same link.
     """
 
     def __init__(self,
@@ -263,7 +259,6 @@ class Signpost:
         :param link: Optional origin :class:`Link` header (not parsed further) for further attributes
 
         :raise ValueError: If a plain string value is invalid for the corresponding type-checked classes :class:`LinkRel`, :class:`AbsoluteURI` or :class:`MediaType`,
-
         """
 
         if isinstance(rel, LinkRel):
@@ -324,7 +319,7 @@ class Signpost:
         if self.profiles:
             strs.append('profile="%s"' % " ".join(self.profiles))
         if self.context:
-            strs.append('context="%s"' % self.context)
+            strs.append('anchor="%s"' % self.context)
         return "; ".join(strs)
 
     def _eq_attribs(self) -> Iterable[object]:
@@ -365,22 +360,21 @@ class Signpost:
             h ^= hash(e)
         return h
 
-    def with_context(self, context: Union[AbsoluteURI, str, None]):
+    def with_context(self, context: Union[AbsoluteURI, str, None]) -> Signpost:
         """Create a copy of this signpost, but with the specified context.
         
-        If the context is None, it means the copy will not have a context.
+        If the context is ``None``, it means the copy will not have a context.
         """
         return Signpost(self.rel, self.target, self.type, self.profiles, context, self.link)
-
 
 class Signposting(Iterable[Signpost], Sized):
     """Signposting links for a given resource.
 
     Links are categorized according to `FAIR`_ `signposting`_ conventions and 
-    split into different attributes like `citeAs` or `describedBy`.
+    split into different attributes like :attr:`citeAs` or :attr:`describedBy`.
 
-    It is possible to iterate over this class or use the `signposts` property to
-    find all recognized signposts.
+    It is possible to iterate over this class or use the :attr:`signposts` 
+    property to find all recognized signposts.
 
     Note that in the case of a resource not having any signposts, instances
     of this class are considered false.
@@ -399,8 +393,8 @@ class Signposting(Iterable[Signpost], Sized):
 
     other_contexts: Set[AbsoluteURI]
     """Other resource URLs which signposting has been provided for. 
-    
-    Use :meth:`for_context` to retrieve their signpostings, or filter the full list of signposts from :prop:`signposts` according to :attr:`Signpost.context`
+
+    Use :meth:`for_context` to retrieve their signpostings, or filter the full list of signposts from :attr:`signposts` according to :attr:`Signpost.context`
     """
 
     authors: Set[Signpost]
@@ -409,7 +403,7 @@ class Signposting(Iterable[Signpost], Sized):
     describedBy: Set[Signpost]
     """Metadata resources about this resource and its items, typically in a Linked Data format.
 
-    Resources may require content negotiation, check `Signpost.type` attribute
+    Resources may require content negotiation, check :attr:`Signpost.type` attribute
     (if present) for content type, e.g. ``text/turtle``.
     """
 
@@ -419,20 +413,20 @@ class Signposting(Iterable[Signpost], Sized):
     items: Set[Signpost]
     """Items contained by this resource, e.g. downloads.
 
-    The content type of the download may be available as `Signpost.type` attribute.
+    The content type of the download may be available as :attr:`Signpost.type` attribute.
     """
 
     linksets: Set[Signpost]
     """Linkset resources with further signposting for this resource (and potentially others).
 
-    A `linkset`_ is a JSON or text serialization of Link headers available as a
+    A `Linkset`_ is a JSON or text serialization of Link headers available as a
     separate resource, and may be used to externalize large collection of links, e.g.
-    thousands of "item" relations.
+    thousands of ``item`` relations.
 
     Resources may require content negotiation, check ``Link["type"]`` attribute
     (if present)  for content types ``application/linkset`` or ``application/linkset+json``.
 
-    .. _linkset: https://datatracker.ietf.org/doc/draft-ietf-httpapi-linkset/
+    .. _Linkset: https://datatracker.ietf.org/doc/draft-ietf-httpapi-linkset/
     """
 
     citeAs: Optional[Signpost]
@@ -447,8 +441,9 @@ class Signposting(Iterable[Signpost], Sized):
     def __init__(self, 
                  context_url: Union[AbsoluteURI, str] = None, 
                  signposts: Iterable[Signpost] = None,
-                 include_no_context: bool = True):
-        """Construct a Signposting from a list of :class:`Signpost`s.
+                 include_no_context: bool = True,
+                 warn_duplicate=True):
+        """The constructor takes a an iterable of :class:`Signpost`.
 
         Signposts are filtered by the matching `context_url` (if provided), 
         then assigned to attributes like :attr:`citeAs` or :attr:`describedBy`
@@ -461,16 +456,17 @@ class Signposting(Iterable[Signpost], Sized):
         A Signposting object is equivalent to boolean `False` in conditional expression 
         if it is empty, that is ``len(signposting)==0``, indicating no signposts 
         were discovered for the given context. However the remaining 
-        ``signposts`` will still be available from :attr:`signposts`, as
-        indicated by :attr:`other_contexts`.
+        signposts will still be available from :attr:`signposts`, as
+        indicated by :attr:`other_contexts` and retrievable with :meth:`for_context`.
         
         :param context_url: the resource to select signposting for, or any signposts if ``None``.            
-        :param signposts: An iterable of :class:`Signpost`s that should be considered for selecting signposting.
-        :param include_no_context: If true (default), consider signposts without explicit context, 
+        :param signposts: An iterable of :class:`Signpost` that should be considered for selecting signposting.
+        :param include_no_context: If `True` (default), consider signposts without explicit context, 
             assuming they are about ``context_url``. 
-            If false, such signposts are ignored for assignment, 
+            If `False`, such signposts are ignored for assignment, 
             but remain available from :attr:`signposts`.
-        :raise ValueError: If ``include_no_context`` is false, but ``context_url`` was not provided or None.
+        :param warn_duplicate: If `True` (default), warn of duplicate signposts that can't be assigned.
+        :raise ValueError: If ``include_no_context`` is false, but ``context_url`` was not provided or `None`.
         """
         if not include_no_context and not context_url:
             raise ValueError("Can't exclude signposts without context when not providing context_url; try include_no_context=True")
@@ -509,20 +505,20 @@ class Signposting(Iterable[Signpost], Sized):
                 if context:
                     self.other_contexts.add(context)
             elif s.rel is LinkRel.cite_as:
-                if self.citeAs:
-                    warnings.warn("Ignoring additional cite-as signposts")
+                if self.citeAs and self.citeAs.target != s.target:
+                    warn("Ignoring additional cite-as signposts") if warn_duplicate else None
                     self._extras.add(s)
                 else:
                     self.citeAs = s
             elif s.rel is LinkRel.license:
-                if self.license:
-                    warnings.warn("Ignoring additional license signposts")
+                if self.license and self.license.target != s.target:
+                    warn("Ignoring additional license signposts") if warn_duplicate else None
                     self._extras.add(s)
                 else:
                     self.license = s
             elif s.rel is LinkRel.collection:
-                if self.collection:
-                    warnings.warn("Ignoring additional collection signposts")
+                if self.collection and self.collection.target != s.target:
+                    warn("Ignoring additional collection signposts") if warn_duplicate else None
                     self._extras.add(s)
                 else:
                     self.collection = s
@@ -537,7 +533,7 @@ class Signposting(Iterable[Signpost], Sized):
             elif s.rel is LinkRel.type:
                 self.types.add(s)
             else:
-                warnings.warn("Unrecognized link relation: %s" % s.rel)
+                warn("Unrecognized link relation: %s" % s.rel)
                 # NOTE: This means a new enum member in LinkRel that we should handle above
                 self._extras.add(s)
 
@@ -546,49 +542,63 @@ class Signposting(Iterable[Signpost], Sized):
         """All FAIR Signposts with recognized relation types.
         
         This may include any additional signposts for link relations
-        that only expect a single link, like :prop:`citeAs`, as well
-        as any signposts for other contexts as listed in :prop:`other_contexts`.
-
-        For only the signposts that match the given context, use 
-        :meth:`__iter__`, e.g. ``for s in signposting`` or ``matched = set(signposts)``
+        that only expect a single link, like :attr:`citeAs`, as well
+        as any signposts for other contexts as listed in :attr:`other_contexts`.
         """
         return frozenset(itertools.chain(self, self._others))
 
-    def for_context(self, context_uri:Union[AbsoluteURI, str, None]):
+    def _signposts_with_explicit_context(self) -> Iterable[Signpost]:
+        """Iterate over all our signposts, making context explicit if possible.
+
+        Variant of ::attr:signposts, gives a generator.
+
+        If ::attr:`context_url` is set, then signposts without a context will be
+        excluded (to avoid them leaking across contexts). 
+        """
+        for s in self:
+            if not s.context and self.context_url:
+                # Clone to make implicit context explicit
+                yield s.with_context(self.context_url)
+            else:
+                yield s
+        for o in self._others:
+            if not o.context and self.context_url: 
+                warn("Ignoring signpost with unknown context: %s" % o)
+                continue
+            yield o
+
+    def for_context(self, context_url:Union[AbsoluteURI, str, None]) -> Signposting:
         """Return signposting for given context URI.
         
-        This will select an alternative view of the signposts from :attr:`signposts`
-        filtered by the given ``context_uri``.
+        This will select an alternative view of the :attr:`signposts`
+        filtered by the given ``context_url``.
 
         The remaining signposts and their contexts will be included under 
-        :attr:`Signpost.signposts` -- any signposts with implicit context will
-        be replaced with having an explicit context :attr:`self.context_url`.
+        :attr:`signposts` -- any signposts with implicit context will
+        be replaced with having an explicit context from :attr:`context_url`.
 
-        Tip: To ensure all signposts have explicit context, use 
-        ``s.for_context(s.context_uri)``
+        **Tip**: To ensure all signposts have explicit context, use 
+        ``s.for_context(s.context_url)``
 
-        :param context_uri: The context to select signposts from. 
-            The URI should be a member of :attr:`contexts` or equal to :attr:`context`, 
+        :param context_url: The context to select signposts from. 
+            The URI should be a member of :attr:`other_contexts` or equal to :attr:`context_url`, 
             otherwise the returned Signposting will be empty.
-            If the context_uri is `None`, then the :attr:`Signpost.context` is ignored
-            and any signposts will be considered.
+
+            If this parameter is `None`, then the individual :attr:`Signpost.context` 
+            values are ignored and any signposts will be considered.
         """
-        include_no_context = context_uri is None
+        include_no_context = context_url is None
         if include_no_context:
-            # If context_uri is None, then include any implicit contexts as-is
-            our_signposts = self
+            # include any implicit contexts as-is
+            our_signposts: Iterable[Signpost] = self.signposts
         else:
             # ensure explicit contexts, so they don't get lost
-            our_signposts = (s.with_context(self.context_url) for s in self)
-
-        return Signposting(context_uri, 
-                           # Chain in own signposts, 
-                           # in case they want to call for_context() 
-                           # back to our context.  
-                           itertools.chain(our_signposts, self._others),
+            our_signposts = self._signposts_with_explicit_context()
+        return Signposting(context_url, 
+                           our_signposts,
                            include_no_context=include_no_context)
 
-    def __len__(self):
+    def __len__(self) -> int:
         """Count how many FAIR Signposts were recognized for the given context"""
         # Note: tuple(self) fails here, as tuple will call our __len__ to pre-allocate
         #return len(tuple(self))
@@ -627,6 +637,9 @@ class Signposting(Iterable[Signpost], Sized):
         although each :attr:`Signpost.context` are included when comparing list of signposts. 
         This distinction becomes significant when comparing signposts without explicit
         context, loaded from two different contexts.
+
+        **Tip**: To compare two Signposting's using only explicit ``Signpost.context``s, use
+        ``a.for_context(a.context_url) == b.for_context(b.context_url)``
         """
         if not isinstance(o, Signposting):
             return False
@@ -649,13 +662,12 @@ class Signposting(Iterable[Signpost], Sized):
         ##    h ^= hash(e)
         return h
 
-
-    def _repr_signposts(self, signposts):
+    def _repr_signposts(self, signposts) -> str:
         """String representation of a list of signposts"""
         # This is usually a short list, so no need for max-trimming and ...
         return " ".join(set(d.target for d in signposts))
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         repr = []
         if self.context_url:
             repr.append("context=%s" % self.context_url)
@@ -680,12 +692,138 @@ class Signposting(Iterable[Signpost], Sized):
 
         return "<Signposting %s>" % "\n ".join(repr)
 
-    def __str__(self):
-        """Represent signposts as HTTP Link headers.
+    def __str__(self) -> str:
+        """Represent all :atttr:`signposts` as HTTP Link headers.
         
         Note that these are reconstructed from the recognized link relations only,
-        and do not include unparsed additional link attributes or links with different contexts.
+        and do not include unparsed additional link attributes.
+        
+        Signposts with other contexts are included in this listing.
 
         See also `Signpost.link`
         """
-        return "\n".join(map(str, self))
+        return "\n".join(map(str, self.signposts))
+
+    def __or__(self, other: Signposting) -> Signposting:
+        """Merge two Signposting instances.
+        
+        The context of the new Signposting is determined as:  
+        
+        a) The context of the left-hand Signposting, if explicit; or
+        b) The context of the right-hand Signposting, if explicit; or
+        c) No context
+        
+        When merging Signpost, any implicit contexts are made explicit
+        from their original :attr:`Signposting.context_url` if specified. 
+        
+        If neither Signposting has a context, then the new `Signposting`
+        is constructed with ``include_no_context=True`` meaning that only
+        signposts _without_ context are considered. Otherwise only
+        signpost _with_ the determined context are considered. 
+        
+        **Tip**: To adapt signposts without context, use :meth:`Signposting.__add__` instead
+
+        If multiple signposts match singular properties like :attr:`citeAs` but with 
+        different targets, Signpost from this instance (left-hand) 
+        will be preferred after merging, however all signposts will be 
+        included in the iteration over the returned ``Signposting``. 
+
+        Duplicate Signpost (as determined by :meth:`Signpost.__eq__`) 
+        will be ignored, it is unspecified which Signpost will be rejected 
+        (this should primarily affects :attr:`Signpost.link`).
+
+        As an example, if the left-hand Signpost ``a`` with context ``http://example.com/doc/1`` had::
+
+            Link <http://example.com/pid/A>;rel=cite-as
+            Link <http://example.com/author/1>;rel=author``
+            
+        and the right-hand ``b`` with context ``http://example.com/doc/2`` had::
+
+            Link <http://example.com/pid/B>;rel=cite-as"
+            Link <http://example.com/author/2>;rel=author;context="http://example.com/doc/1`
+
+        then the resulting Signposting ``a|b`` would contain::
+
+            Link <http://example.com/pid/A;rel=cite-as;context="http://example.com/doc/1"
+            Link <http://example.com/author/1>;rel=author;context="http://example.com/doc/1"
+            Link <http://example.com/author/2>;rel=author;context="http://example.com/doc/1"
+
+        In this case ``…pid/B`` is ignored in the merged signposting as it relates to ``…doc/2`` 
+        and not the determined context ``http://example.com/doc/1``, which on the other hand
+        has been made explicit in all its direct signposts.
+
+        The complete set of merged signposts (regardless of their context) 
+        is available in :attr:`Signposting.signposts` in the returned instance.
+
+        :param other: Another `Signposting` instance which signposts are to be merged with ours
+        :return: A new `Signposting` instance from the merged list of signposts. 
+        :raise TypeError: If `other` is not an instance of `Signposting`
+        """
+        if not isinstance(other, Signposting):
+            raise TypeError("Can only merge with Signposting instances, not: " % type(other))
+        # Decide if the merged Signposting will have a context. 
+        # Left hand has preference.
+        newContext = self.context_url or other.context_url or None
+        if newContext:
+            # Merge with explicit contexts so that Signposts can be compared
+            merged: Iterable[Signpost] = itertools.chain(self._signposts_with_explicit_context(), 
+                    other._signposts_with_explicit_context())
+        else:
+            # Both are context-free, merge them as-is, but prefer self
+            merged = itertools.chain(self.signposts,other.signposts)
+        return Signposting(newContext, merged,                            
+                           include_no_context=not newContext,
+                           warn_duplicate=False)
+
+    def __add__(self, other: Union[Signposting, Iterable[Signpost]]) -> Signposting:
+        """Create a merged Signposting by overriding new Signposts from another.
+
+        The returned Signposting instance will have the same context as this instance.
+
+        `other` is considered an iterable of `Signpost`s -- if it is a `Signposting` instance,
+        this method will iterate over its direct signposts only if its context is ``None`` or matches
+        the current context, otherwise it will select the current context using :meth:`Signposting.for_context`.
+        
+        If the added Signpost's context is `None` or match the current :attr:`context_url`` 
+        they will __replace__ or append the existing signposts from this instance. 
+        
+        For instance, if the left-hand Signpost ``a`` had::
+
+            Link <http://example.com/pid/A>;rel=cite-as
+            Link <http://example.com/author/1>;rel=author``
+            
+        and the right-hand ``b`` had::
+
+            Link <http://example.com/pid/B>;rel=cite-as
+            Link <http://example.com/author/2>;rel=author
+
+        then the resulting Signposting ``a+b`` would contain::
+
+            Link <http://example.com/pid/B>;rel=cite-as
+            Link <http://example.com/author/1>;rel=author``
+            Link <http://example.com/author/2>;rel=author
+
+        (Note: iterating over the returned `Signposting` would include the relegated ``…pid/A``. 
+        Take care of operatand order if adding multiple Signpostings).
+        
+        Added signposts with other contexts will be ignored and
+        __not__ added to the resulting `signposts`, however existing 
+        signposts from other contexts in this instance are preserved.
+
+        To do a full merge across contexts, use instead ``a | b``, see :meth:`__or__`
+
+        :param other: Either an `Iterable` (`Set`, `List`, etc) of `Signpost` instances, or another `Signposting` instance.
+            The signposts are to be added to our signposts, if matching the determined context.
+        :return: A new `Signposting` instance from the summed list of signposts. 
+            Signposts from `other` may overridde signposts from this instance.
+        :raise TypeError: If `other` is not an instance of `Signposting` or an iterable of Signposts.
+        """
+        if (isinstance(other, Signposting) and 
+                self.context_url and other.context_url and 
+                other.context_url != self.context_url):
+            to_add: Iterable[Signpost] = other.for_context(self.context_url)
+        else:
+            to_add = (s for s in other if s.context == self.context_url or not s.context)
+        return Signposting(self.context_url, itertools.chain(to_add, self.signposts), 
+            # NOTE: We chain the added ones first so they can override the singular properties like citeAs
+                include_no_context=True, warn_duplicate=False)